+++ src/weather/formatter.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use crate::config::Config;
 use crate::location::Location;
-use crate::weather::Weather;
+use crate::weather::{Convert, Weather};
 
 macro_rules! tag {
     ($tags:expr, $option:expr) => {
@@ -41,36 +41,25 @@ impl<'a> Formatter<'a> {
     fn tags(&self) -> HashMap<String, String> {
         let mut tags = HashMap::new();
 
-        let city = Some(&self.location.city);
-        let icon = Some(self.icon());
-        let temperature = Some(self.weather.temp().convert(&self.config.units.temperature));
-        let temperature_feels_like = self
-            .weather
-            .temp_feels_like()
-            .map(|kelvin| kelvin.convert(&self.config.units.temperature));
-        let temperature_max = self
-            .weather
-            .temp_max()
-            .map(|kelvin| kelvin.convert(&self.config.units.temperature));
-        let temperature_min = self
-            .weather
-            .temp_min()
-            .map(|kelvin| kelvin.convert(&self.config.units.temperature));
-        let pressure = self.weather.pressure();
+        let city = &self.location.city;
+        let icon = self.icon();
+
+        let temperature = self.convert(self.weather.temp());
+        let temperature_feels_like = self.convert(self.weather.temp_feels_like());
+        let temperature_max = self.convert(self.weather.temp_max());
+        let temperature_min = self.convert(self.weather.temp_min());
+        let dew_point = self.convert(self.weather.dew_point());
+
+        let precipitation = self.convert(self.weather.precipitation());
+        let precipitation_chance = self.weather.precipitation_chance();
+        let clouds = self.weather.clouds();
         let humidity = self.weather.humidity();
-        let wind_speed = self
-            .weather
-            .wind_speed()
-            .map(|ms| ms.convert(&self.config.units.wind_speed));
-        let cloud_percentage = self.weather.clouds();
-        let visibility = self
-            .weather
-            .visibility()
-            .map(|visibility| visibility.convert(&self.config.units.distance));
-        let precipitation = self
-            .weather
-            .precipitation()
-            .map(|precipitation| precipitation.convert(&self.config.units.precipitation));
+        let visibility = self.convert(self.weather.visibility());
+        let wind_speed = self.convert(self.weather.wind_speed());
+        let pressure = self.weather.pressure();
+        let uv_index = self.weather.uvi();
+        let air_quality_index = self.weather.aqi();
+
         let sunrise = self.weather.sunrise().map(|time| time.format("%H:%M"));
         let sunset = self.weather.sunset().map(|time| time.format("%H:%M"));
 
@@ -80,23 +69,33 @@ impl<'a> Formatter<'a> {
         tag!(tags, temperature_feels_like);
         tag!(tags, temperature_max);
         tag!(tags, temperature_min);
-        tag!(tags, pressure);
-        tag!(tags, humidity);
-        tag!(tags, wind_speed);
-        tag!(tags, cloud_percentage);
-        tag!(tags, visibility);
+        tag!(tags, dew_point);
         tag!(tags, precipitation);
+        tag!(tags, precipitation_chance);
+        tag!(tags, clouds);
+        tag!(tags, humidity);
+        tag!(tags, visibility);
+        tag!(tags, wind_speed);
+        tag!(tags, pressure);
+        tag!(tags, uv_index);
+        tag!(tags, air_quality_index);
         tag!(tags, sunrise);
         tag!(tags, sunset);
         tags
     }
 
-    fn icon(&self) -> String {
-        let condition = self.weather.weather_condition().to_string();
-        self.config
+    fn icon(&self) -> Option<String> {
+        let condition = self.weather.weather_condition()?.to_string();
+        let icon = self
+            .config
             .icons
             .get(&condition)
             .unwrap_or(&condition)
-            .to_string()
+            .to_string();
+        Some(icon)
+    }
+
+    fn convert<T: Convert>(&self, option: Option<T>) -> Option<String> {
+        option.map(|value| value.convert(&self.config.units))
     }
 }